chore!: expose board property of game-board
BREAKING CHANGE: in order for ease of access to the board. The previous
way to accessing it was via a getter which returned a copy of the board.
However, this copy was only shallow, allowing the nested Ship objects
to still be modified, defeating the purpose of returning a copy.

A deep copy was not feasible since window.structuredClone does not
support cloning functions, namely, Ship's functions.

Ultimately, I decided on simply making the board variable public.